// web/src/app/shared/api-client/model/exercise-dto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ExerciseFeedbackDto } from './exercise-feedback-dto';
import { ExerciseDtoContent } from './exercise-dto-content';
import { SubmitAnswerRequest } from './submit-answer-request';
import { ExerciseDtoProperties } from './exercise-dto-properties';


export interface ExerciseDto { 
    id: string;
    type: string;
    content?: ExerciseDtoContent;
    input?: SubmitAnswerRequest;
    feedback?: ExerciseFeedbackDto;
    properties?: ExerciseDtoProperties;
}